@@ -1,20 +1,20 @@
 {
-  "homepage": "https://midhunreal1.github.io/react-portfolio",
+  "homepage": "https://midhunreal1.github.io/portfolio-website",
   "name": "react-portfolio-website-1",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "emailjs-com": "^3.2.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-icons": "^4.6.0",
-    "react-scripts": "5.0.1",
-    "react-scroll": "^1.8.9",
-    "swiper": "^8.4.4",
-    "web-vitals": "^2.1.4"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-icons": "^5.2.1",
+    "react-scripts": "^5.0.1",
+    "react-scroll": "^1.9.0",
+    "swiper": "^11.1.9",
+    "web-vitals": "^4.2.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -43,6 +43,6 @@
     ]
   },
   "devDependencies": {
-    "gh-pages": "^4.0.0"
+    "gh-pages": "^6.1.1"
   }
 }
