@@ -1,5 +1,11 @@
-import ReactDOM from 'react-dom'
-import App from './App'
-import './index.css'
+import React from 'react';
+import {createRoot} from 'react-dom/client';  
+import App from './App';
+import './index.css';
 
-ReactDOM.render(<App/>,document.querySelector("#root"))
+const root = createRoot(document.getElementById('root'));
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
